+++ Ray Tracer/Sphere.h
@@ -12,7 +12,7 @@ class Sphere : public SceneObject
 {
 public:
 	Sphere();
-	Sphere(int id, Vector3D center, double radius, Color color, float reflectionIndex);
+	Sphere(int id, Vector3D center, double radius, Color color, double reflectionIndex);
 
 	bool getClosestIntersection(Ray ray, Vector3D& intersectionPoint, double& intersectionDistance) override;
 	Vector3D getNormal(Vector3D pointOnObject) override;
